Replace deprecated CGDisplayCurrentMode with CGDisplayCopyDisplayMode.
Display verbose display information as text rather than as an XML/NeXT-style plist dump.

This means we now require 10.6+; change the compilation instructions to make this clear.

Note that the O3 method doesn't compile against 10.7 64-bit or 10.8+ at *all*.  The IOService method is deprecated as of 10.9, so it should probably last a bit longer.